Move markdown text formatting code to separated file
