+++ lib/widgets/markdown_text.dart
@@ -27,3 +27,43 @@
 // SOFTWARE.
 ///
 /// Authors: Tony Chen
+
+import 'package:flutter/material.dart';
+import 'package:flutter_markdown/flutter_markdown.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+import 'package:markdown_widgets/constants/constants.dart'
+    show contentWidthFactor;
+
+class MarkdownText extends StatelessWidget {
+  final String data;
+
+  const MarkdownText({Key? key, required this.data}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: FractionallySizedBox(
+        widthFactor: contentWidthFactor,
+        child: MarkdownBody(
+          data: data,
+          onTapLink: (text, href, title) async {
+            if (href != null) {
+              final uri = Uri.parse(href);
+              if (await canLaunchUrl(uri)) {
+                await launchUrl(uri);
+              } else {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  SnackBar(content: Text('Could not launch $href')),
+                );
+              }
+            }
+          },
+          styleSheet: MarkdownStyleSheet(
+            p: const TextStyle(fontSize: 16),
+          ),
+        ),
+      ),
+    );
+  }
+}
